+++ c++/src/QRAlgorithms.cpp
@@ -16,8 +16,8 @@ namespace bms
     G_.reserve(kmax);
     for (DenseIndex i = 0; i < kmax; ++i)
     {
-      double c = 1 / sqrt(i + 2.);
-      G_.emplace_back(i, c, sqrt(i + 1.)*c);
+      double c = 1 / sqrt(double(i) + 2.);
+      G_.emplace_back(i, c, sqrt(double(i) + 1.)*c);
     }
   }
 }
